requirements.txt
customtkinter>=5.2.2
Pillow>=10.4.0
pefile>=2024.8.26
psutil>=6.1.0
pystray>=0.19.5
watchdog>=6.0.0
